refactor: [1-11] model package -> domain package
