+++ src/main/java/racingcar/domain/CarRecords.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import java.util.ArrayList;
 import java.util.Iterator;
