+++ src/main/java/racingcar/domain/RaceResult.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import java.util.HashMap;
 import java.util.Iterator;
