+++ src/test/java/racingcar/CarNamesTest.java
@@ -3,7 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.model.CarNames;
+import racingcar.domain.CarNames;
 
 import java.util.List;
