+++ src/test/java/racingcar/CarsTest.java
@@ -5,9 +5,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.model.CarNames;
-import racingcar.model.Cars;
-import racingcar.model.MoveStrategy;
+import racingcar.domain.CarNames;
+import racingcar.domain.Cars;
+import racingcar.domain.MoveStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
